feat: move validation to client side
